requirements.txt
python-telegram-bot>=21.7
playwright==1.48.0
aiohttp==3.10.5
python-dotenv==1.0.1
aiosqlite==0.20.0
pytz==2024.1
pillow==10.4.0
psutil==6.0.0
beautifulsoup4==4.12.3
google-generativeai>=0.4.0

